move HeatProduction dataclass from heat2018 to common
