+++ src/climatevision/generator/common/dataclasses.py
@@ -3,7 +3,7 @@
 from dataclasses import dataclass
 from typing import Any
 
-from ...utils import div
+from ..utils import div
 
 
 @dataclass(kw_only=True)
